Refactor api.py: extract Activity / Audit / Hook-Feed routes
Continues the api.py split. Stacks on top of #370.

- /audit + /audit/costs (2 endpoints)
- /hooks + /activity/hooks + /activity/active (3 endpoints)
- /activity + /activity/stats (2 endpoints)
- /activity/stream SSE feed (1 endpoint)

Total: 8 endpoints moved → routes/activity.py.

Three small surfaces that all share `audit`, `hooks`, and `activity`
as dependencies — bundled together because they conceptually all
expose "what's happening" in the system.

api.py: 7,235 → 7,151 lines (-84, -1.2%)
Cumulative since main: 11,138 → 7,151 (-3,987 lines, -35.8%)

Tests: 1542 passed, 1 skipped (federation tests need nacl, pre-existing).
